Структурировал код, раскидав функции по модулям
